Implementation of Processing a Transaction, with respective tests endpoints and errors

+++ src/main/java/com/unq/dapp_grupo_e/utilities/factories/UserFactory.java
@@ -33,7 +33,7 @@ public class UserFactory {
         return user;
     }
 
-    public static User createWithIdDataAndOperations(Long idUser, String name, String email,
+    public static User createWithIdDataAndOperations(Integer idUser, String name, String email,
                                                     Integer operationsSetted, Integer reputationPoints) {
         var user = new User();
         user.setIdUser(idUser);
